+++ src/Nono.Engine/DiffLine.cs
@@ -11,9 +11,9 @@ namespace Nono.Engine
 
         public LineIndex Index { get; }
 
-        public IEnumerable<int> NonEmptyIndexes()
+        public IEnumerable<ushort> NonEmptyIndexes()
         {
-            for (int i = 0; i < Length; i++)
+            for (ushort i = 0; i < Length; i++)
             {
                 if (this[i] != Box.Empty)
                     yield return i;
