+++ tests/Nono.Engine.Tests/CombinationsShould.cs
@@ -16,10 +16,10 @@ namespace Nono.Engine.Tests
 
 
         [Theory]
-        [InlineData(new uint[] {10}, 12, 3L)]
-        [InlineData(new uint[] { 2},  4, 3L)]
-        [InlineData(new uint[] { 1},  3, 3L)]
-        public void OneBlockCount(uint[] task, int length, long expectedCount)
+        [InlineData(new ushort[] {10}, 12, 3L)]
+        [InlineData(new ushort[] { 2},  4, 3L)]
+        [InlineData(new ushort[] { 1},  3, 3L)]
+        public void OneBlockCount(ushort[] task, int length, long expectedCount)
         {
             var count = Combinations.Count(task, length);
 
@@ -27,10 +27,10 @@ namespace Nono.Engine.Tests
         }
 
         [Theory]
-        [InlineData(new uint[] {2, 2}, 6, 3L)]
-        [InlineData(new uint[] {2, 1}, 5, 3L)]
-        [InlineData(new uint[] {1, 1}, 4, 3L)]
-        public void TwoBlocksCount(uint[] task, int length, long expectedCount)
+        [InlineData(new ushort[] {2, 2}, 6, 3L)]
+        [InlineData(new ushort[] {2, 1}, 5, 3L)]
+        [InlineData(new ushort[] {1, 1}, 4, 3L)]
+        public void TwoBlocksCount(ushort[] task, int length, long expectedCount)
         {
             var count = Combinations.Count(task, length);
 
@@ -38,8 +38,8 @@ namespace Nono.Engine.Tests
         }
 
         [Theory]
-        [InlineData(new uint[] {6, 2, 16, 1}, 35, 330L)]
-        public void CountComplex(uint[] task, int length, long expectedCount)
+        [InlineData(new ushort[] {6, 2, 16, 1}, 35, 330L)]
+        public void CountComplex(ushort[] task, int length, long expectedCount)
         {
             var count = Combinations.Count(task, length);
